design: 제작한 프로젝트 하이파이 구현 #140
